Python_lib/boomerang/requirements.txt
alabaster==0.7.10
astroid==2.0.4
attrs==17.4.0
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
coverage==4.5
dataclasses==0.6
docutils==0.14
elasticsearch==6.3.1
idna==2.6
imagesize==0.7.1
isort==4.3.4
Jinja2==2.10
jsonpickle==0.9.6
kafka-python==1.4.1
lazy-object-proxy==1.3.1
lockfile==0.12.2
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
pbr==3.1.1
pluggy==0.6.0
PTable==0.9.2
py==1.5.2
Pygments==2.2.0
pylint==2.1.1
pytest==3.4.0
pytest-cov==2.5.1
pytest-pythonpath==0.7.2
python-daemon==2.1.2
python-dateutil==2.8.0
python-interface==1.4.0
pytz==2017.3
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.7
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
tornado==4.5.3
typed-ast==1.1.0
urllib3==1.22
wrapt==1.10.11
